# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /user/{phone}:
    get:
      consumes:
      - application/json
      description: با استفاده از شماره تلفن، جزئیات کاربر را دریافت می‌کند
      parameters:
      - description: شماره تلفن کاربر
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: کاربر یافت نشد
          schema:
            additionalProperties: true
            type: object
        "500":
          description: خطا در دریافت کاربر
          schema:
            additionalProperties: true
            type: object
      summary: دریافت جزئیات یک کاربر
      tags:
      - Users
...
swagger: "2.0"